collect_service/requirements.txt
asgiref==3.11.1
attrs==26.1.0
black==26.3.1
blinker==1.9.0
click==8.3.1
colorama==0.4.6
Django==6.0.3
django-cors-headers==4.9.0
djangorestframework==3.17.0
drf-spectacular==0.29.0
Flask==3.1.3
Flask-SQLAlchemy==3.1.1
greenlet==3.3.2
inflection==0.5.1
isort==8.0.1
itsdangerous==2.2.0
Jinja2==3.1.6
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
MarkupSafe==3.0.3
mypy_extensions==1.1.0
numpy==2.4.3
opencv-python==4.13.0.92
packaging==26.0
pathspec==1.0.4
pillow==12.1.1
platformdirs==4.9.4
pytokens==0.4.1
PyYAML==6.0.3
referencing==0.37.0
rpds-py==0.30.0
SQLAlchemy==2.0.48
sqlparse==0.5.5
typing_extensions==4.15.0
tzdata==2025.3
uritemplate==4.2.0
Werkzeug==3.1.6
